attempting to fix memory allocation issue - Incommplete

+++ watchmal/dataset/cnn/cnn_dataset.py
@@ -50,7 +50,7 @@ class CNNDataset(H5Dataset):
         """
         super().__init__(h5file, use_memmap)
         
-        self.pmt_positions = np.load(pmt_positions_file)['pmt_image_positions']
+        self.pmt_positions = np.load(pmt_positions_file)['pmt_image_positions'].astype(int)
         self.use_times = use_times
         self.use_charges = use_charges
         self.data_size = np.max(self.pmt_positions, axis=0) + 1
